requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
decorator==5.1.1
Django==4.1.2
djangorestframework==3.14.0
executing==1.1.1
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
toml==0.10.2
traitlets==5.5.0
wcwidth==0.2.5
